Output individual posts to unused .posts directory
tAMG doesn't use individual post pages, but these are required by
Jekyll. The previous solution repeatedly clobbered a single file
(.trash.html), but this raised Jekyll errors. This changes that solution
to output to a similarly hidden directory (.posts) with individual files
under this, which generates no Jekyll errors. Since the site is uploaded
with rsync, which ignores dotfiles unless told not to, this change
avoids Jekyll errors while still allowing simple syncing of the site.